Updates to handle fetching responsive / screen-resized images.
Modified the transfer of attributes from the target link to allow adding 'srcset' and 'sizes' attributes
(or any other - configurable via the importAttrs option).

Fixed up the allowedTypes validation to handle having non-image links alongside the gallery ones.
This is now done as a pre-validation step with tagging of the invalid links to exclude them from being added to navigation.